requirement.txt
celery
gunicorn
gevent
django==1.6.5
django-celery
south
requests
tornado==3.2
flower
redis
mailer